requirements.txt
aiogram==3.21.0
sqlalchemy>=2.0.25
alembic>=1.13.1
python-dotenv==1.0.0
psycopg2-binary>=2.9.10
supabase>=2.18.1 